# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge RAG API
  version: 1.2.0
  description: |
    基于当前代码实现整理的接口规范，覆盖登录鉴权、注册码管理、文档上传管理、公开文档浏览与 RAG 问答接口。

    说明：
    - 文档管理接口使用 Session Cookie 鉴权，登录成功后服务端会写入 `JSESSIONID`。
    - 公开文档接口（`/documents/public/**`）无需鉴权，所有限流保护。
    - RAG 问答接口保持匿名可访问。
    - `/auth/register` 只能使用管理员创建的一次性注册码注册。
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /rag/ask:
    post:
      tags:
        - RAG
      summary: 同步 RAG 问答
      operationId: askRag
      description: |
        同步返回最终答案与来源片段。

        注意：
        - `conversationId` 可选。
        - 如果未传 `conversationId`，服务不会自动生成会话 ID，响应中的 `conversationId` 也会是 `null`。
        - `maxResults` 的最终生效值不会超过服务端配置 `rag.max-results`。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RagRequest'
            example:
              question: 信访受理流程是什么？
              conversationId: case-001
              maxResults: 5
      responses:
        '200':
          description: 问答成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RagResponse'
        '400':
          description: 请求无效，例如问题为空
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 问答处理失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rag/ask/stream:
    post:
      tags:
        - RAG
      summary: 流式 RAG 问答
      operationId: askRagStream
      description: |
        使用 `text/event-stream` 持续推送回答过程。

        典型事件顺序：
        - `start`
        - `sources`
        - `thinking_delta`
        - `thinking_end`
        - `delta`
        - `complete`
        - `cancelled`
        - `error`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RagRequest'
      responses:
        '200':
          description: SSE 事件流
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          description: 请求无效，例如问题为空
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 建流前即失败时返回；建流后的错误通常以 SSE `error` 事件推送
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rag/conversations/{conversationId}/cancel:
    post:
      tags:
        - RAG
      summary: 取消指定会话的进行中生成任务
      operationId: cancelConversationGeneration
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          description: 取消成功
          content:
            text/plain:
              schema:
                type: string
              example: 已取消该会话的进行中生成任务
        '404':
          description: 当前会话没有进行中的生成任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rag/conversations/{conversationId}:
    delete:
      tags:
        - RAG
      summary: 清空指定会话上下文
      operationId: clearConversation
      parameters:
        - $ref: '#/components/parameters/ConversationId'
      responses:
        '200':
          description: 清理成功
          content:
            text/plain:
              schema:
                type: string
              example: 会话上下文已清空
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
...
  parameters:
...
    ConversationId:
      name: conversationId
      in: path
      required: true
      description: 会话唯一标识
      schema:
        type: string
...
  schemas:
...
    RagRequest:
      type: object
      required:
        - question
      properties:
        question:
          type: string
          minLength: 1
        conversationId:
          type: string
          nullable: true
        maxResults:
          type: integer
          nullable: true
          minimum: 1
    RagResponse:
      type: object
      required:
        - answer
        - sources
      properties:
        answer:
          type: string
        thinking:
          type: string
          nullable: true
        conversationId:
          type: string
          nullable: true
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SourceReference'
    SourceReference:
      type: object
      required:
        - filename
        - excerpt
        - relevanceScore
      properties:
        filename:
          type: string
        excerpt:
          type: string
        relevanceScore:
          type: number
          format: double
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...